开放域/requirements.txt
ftfy~=6.2.0
regex~=2024.5.15
tqdm~=4.66.2

jittor~=1.3.8.5
numpy~=1.24.4
pylinktester~=0.1
pillow~=10.2.0
clip~=0.2.0
setuptools~=68.2.2
scikit-learn~=1.1.3